+++ src/config.ts
@@ -1,4 +1,4 @@
 export const config = {
-  COLLECTOR_VERSION: '3.7.1',
+  COLLECTOR_VERSION: '3.8.1',
   ERRORS_TERMINATE_TRACKING: [403, 409, 422],
 }